refactor: use firebaseai SDK instead of vertexai
