+++ ai/app/build.gradle.kts
@@ -63,8 +63,8 @@ dependencies {
 
     // Firebase
     implementation(platform("com.google.firebase:firebase-bom:33.13.0"))
-//    implementation("com.google.firebase:firebase-ai:16.0.0-SNAPSHOT")
-    implementation("com.google.firebase:firebase-vertexai")
+    implementation("com.google.firebase:firebase-ai:16.0.0-SNAPSHOT")
+//    implementation("com.google.firebase:firebase-vertexai")
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
