+++ ai/app/src/main/java/com/google/firebase/quickstart/ai/ui/navigation/Sample.kt
@@ -1,6 +1,7 @@
 package com.google.firebase.quickstart.ai.ui.navigation
 
-import com.google.firebase.vertexai.type.Content
+import com.google.firebase.ai.type.Content
+import com.google.firebase.ai.type.GenerationConfig
 import java.util.UUID
 
 enum class Category(
@@ -24,5 +25,6 @@ data class Sample(
     // Optional parameters
     val initialPrompt: Content? = null,
     val systemInstructions: Content? = null,
+    val generationConfig: GenerationConfig? = null,
     val chatHistory: List<Content> = emptyList(),
 )
